src/test/java/com/app/dir/domain/EntryTest.java
package com.app.dir.domain;

import static org.junit.Assert.assertEquals;

import java.io.File;
import java.net.URL;

import javax.xml.bind.JAXBContext;
import javax.xml.bind.JAXBException;
import javax.xml.bind.Unmarshaller;

import org.junit.Test;

import com.app.dir.domain.Entry;

public class EntryTest {

	@Test
	public void unmarshallCompanyXml_DummyCompanyXML_PojoWithXmlElements()
			throws JAXBException {
		Entry expected = new Entry();
		expected.setKey("domain");
		expected.setValue("mydomain");


		URL url = this.getClass().getResource("/dummy_entry.xml");
		String filePath = url.getPath();

		File file = new File(filePath);

		JAXBContext jaxbContext = JAXBContext.newInstance(Entry.class);

		// TODO: check xml escape characters. If necessary make fix for those
		// special cases.
		Unmarshaller jaxbUnmarshaller = jaxbContext.createUnmarshaller();

		Entry actual = (Entry) jaxbUnmarshaller.unmarshal(file);

		assertEquals(
				"Unmarshall test failed. Input XML unsuccessfully unmarshall to POJO",
				expected, actual);
	}

}
